Big update
Adjusted MovieDisplay so it fetches all movies into the list along with price and ID

SeatDisplay fetches booking data and compares to current movie selected to show which seats are available and which are occupied.

App.jsx now renders bookings for the first movie selected

Filled the DB with more bookings.

Changed the name of DBcrud to crud and added crud functions to fetch and manipulated data for future admin use aswell.